requirements.txt
whispercpp @ git+https://github.com/aarnphm/whispercpp.git@9bf9a28
llama-cpp-python==0.1.74
pyinstaller==5.13.0
PyQt5==5.15.9